DUPLO-9973 Ensure cached K8s credentials are still valid with ping (#21)
## Change description

Adding ping of Kubernetes cluster to ensure the cached credentials are still valid for a `duplo-jit k8s` call separate from reading the expiration date.

## Type of change
- [x] Bug fix (fixes an issue)
- [ ] New feature (adds functionality)

## Related issues

> Fix [DUPLO-9973](https://app.clickup.com/t/8655600/DUPLO-9973) - Some users report cache issues.

## Additional Notes

Now using the appropriate namespace. Fixed build issues with upgrades to Go versions and GitHub action versions in the workflow configs. 1.17 -> 1.21

Added the logic to get the tenant name from the tenant ID. If necessary, I could add an optimization to cache the tenant flag as the tenant name before it is converted to a tenant ID. It would save a call to the Duplo API, but there's no benefit of course if the ID was provided in the flag in the first place. I am biasing toward simplicity for the time being on that one.

Of note, I also moved the creation of the cache key for k8s creds on tenants to be performed after the name has been determined to so that the cache key for a tenant on k8s will consistently be based on the host and tenant name.